refactor: delete-festival-page 컴포넌트 가독성 좋게 분리
